requirements.txt
# requirements.txt

praw>=7.8,<8.0
tweepy>=4.16,<5.0
requests>=2.32,<3.0
python-dotenv>=1.1,<2.0